build: add raspberry pi tools to debian builder
Currently the raspberry pi compiler and sysroot are a git submodule
being pulled in at the start of every build. This is adding 5+ minutes
to the start of every build. This script will install the raspberry pi
tools into the debian builder so the download does not take place during
every build.