feat(elixir): tcp transport
This adds a TCP transport for establishing incoming and outgoing
TCP connections. In a follow-up, we will need to attach the outgoing
sockets to :ranch so that we can use the same ranch handler for incoming
and outgoing connections. Additionally that future improvement will let
us easily get connection counts from ranch for telemetry purposes.