feat(elixir): add initial encode/decode implementation
Adds tests for encode/decode.

Adds return route to Registry.
Now that we have a registered address for each incoming tcp connection,
this commit adds to it on the return route.

A follow-on that makes use of this code must replace the return route
when it leaves elixir with a %TCPAddress{}, and also turn this handler
into an Ockam.Worker, including implementing `handle_message/2`.

Removed the original wire protocol.