feat(rust): expose onward route information to user workers
So far, when receiving a message, a worker was only able to query the
return route of a message, but not the onward route. For most
messages this route will be empty, however there are cases where a
worker will have to relay a message to another worker in the route.