feat(rust): make context receive retry receiving if wrong type
Previously this responsibility was placed on the user, while also
dropping messages that were deemed "irrelevant". This change makes
sure to repoll the mailbox until a message of the requested type can
be found.

This does create a second loop point in the receive logic which we
will eventually have to resolve with a better mailbox datastructure.