feat(elixir): refactor stream service, add bidirectional client
Support bi-directional stream exchange using remote workers.

This is a first stage to demo with rust, with minimal changes on the
rust side.

Refactor module structure for stream workers

Make index in the stream index storage to be the next index
to fetch from instead of the last message index

Document bidirectional stream modules and examples