feat(rust): implement graceful stop mechanism in ockam node
This commit adds a graceful shutdown mechanism as outlined in #2007
and #2088.  Two primary shutdown strategies exist:

* Immediate      -- legacy behaviour
* Graceful(secs) -- wait a number of seconds while all systems
                    de-allocate.  Failover into Immediate strategy

This commit also exposes the concept of "worker clusters" via the
Context API.  A set of workers (cluster) will be de-allocated
together, in the reverse order in which clusters were first
introduced.

This refactor also reduces the differences between worker relays and
processor relays.  Currently these mechanisms are only used when
shutting down a node, not when trying to stop individual workers or
worker clusters.