fix(rust): fix crash crash scenario on fallback shutdown strategy
This commit fixes a long-standing bug that would crash the executor
during shutdown.  This was the case when the router had already freed
all workers and afterwards ran into a pipe I/O failure.

This patch adds another switch to the router state ("killed") and
gracefully swallows errors when in this state.

This un-breaks the `stop_now()` function on the node context.