build: rust in release mode when invoke from a parent gradle build
When elixir NIFs are built in ockam_vault_software, the cmake build
looks for rust library files in target/release. This change ensures that
when Rust gradle build is invoked from a parent build (like elixir gradle build)
it builds in release mode.

We may need a better way to communicate this intent in a future change.