chore: ensure we checkout to sha when tagging
To tag crates, we use the version and name specified
in `Cargo.toml`, specifying a commit SHA but not checkout to the commit
could give wrong tag details as the source of truth is in Cargo.toml.
We should ensure the Cargo.toml details is that of the specified commit.