feat(rust): implement worker ready status api
A problem we repeatedly run into is that initialisation code of one
worker takes longer than another worker is waiting for.  This is a
common problem in multi-processor environments and one that is not
easy to solve.

This commit implements a simple API where workers and processors
notify the router after running their initialise code which can be
waited on by other workers.