feat(elixir): add prometheus metrics endpoint
This exposes a http /metrics endpoint that prometheus can
scrape.  Metrics exposed are the same than the ones the existing
influxdb exporter produce.

The (http)port where to expose metrics is read from a
`PROMETHEUS_PORT` env variable.   If the variable is
not set,  the prometheus exporter won't be started