test(rust): fix flaky transport tests
Transport tests that were using a random port to create the "listener" sides were
failing when the selected port was already used by another process. These changes
add a loop to retry binding a transport until it succeeds.

It also adds a small change to the `test` macro so that we don't have to explicitly
stop the context at the end of every test.