fix(elixir): fix expected return on channel setup
The return from both
        Ockam.SecureChannel.KeyEstablishmentProtocol.XX.Initiator.setup/2
        Ockam.SecureChannel.KeyEstablishmentProtocol.XX.Responder.setup/2

is a 4-element tuple containing the next events to run.

The `with` statement was failing to match at that point, and was
just returning the non-matched value.

This (I think accidentally) *did* worked as:
        * The remaining clauses of the `with` statement are basically noop
          right now
        * The body of the `with` statement only contains instrumentation code

So for practical purposes the only visible thing at this point was the
instrumentation not being excecuted.