fix(rust): fix tcp router race condition
As described in issue #2730, when a new TCP connection is accepted and
the new worker pair started it may happen that before the subsequent
router registration finishes the worker pair may have already read the
first message from the socket and when resolving the route, the router
may in turn auto-connect to the address and will then make another
registration attempt, ultimately causing an error as there can be only
one mapping for a given address. See issue #2730 for details.

This commit fixes the issue by deferring the spawning of the worker
pair until the registration was completed.