refactor(rust): add an async_drop mechanism for bare context drop
The current implementation of bare Context Drop requires a future
blocking call in a sync function, which causes issues in some of our
tests (and is generally not very elegant).

This commit adds a simple wrapper around receiving a message via a
Oneshot channel (which doesn't require await and can thus be run from
a sync function), and then maps it onto an async message send to the
router to de-allocate a specific resource.

The API for this structure is currently still very specific, but we
can easily change it in the future to allow for different use-cases.