fix(rust): hide tracing logging on client-side ockam cli instance
Previously we were spamming the user with irrelevant log messages from
the embedded node in the ockam CLI when querying data from  a remote
node.  On the other hand, we still want a remote node to log into
their log file.

This change adds a hidden marker flag which is set for spawned nodes
that enables tracing logging, and disables it for the client instance
the user is looking at.

This is not an ideal solution, but means that we can focus on the UX
of the command without having to solve every logging problem at once.