feat(rust): implement transport creation via ockam command
Add the ability to create new transports via the node manager
service.  Status about the creation will be relayed to the calling
CLI.  Any error will be displayed to the user, and new transports
tracked via the node manager.

Currently it's not yet possible to delete a transport, but the route
to the newly created transport is displayed.