feat(elixir): minicbor compatibility and helpers
Implements a minicbor' compatible encoding/decoding layer on top of CBOR
library.
On the encoding side, it works by pre-processing elixir data _before_ encoding
them with CBOR.
On the decoding side,  it post-process CBOR' result into the expected elixir
data type.
On top of this, TypedStruct plugin is provided to help defining structs this way