fix(rust): strip pre-existing `IdentitySecureChannelLocalInfo` entries
This commit introduces a new method:

    pub fn mark(
        mut local_info: Vec<LocalInfo>,
        their_identity_id: IdentityIdentifier,
    ) -> Result<Vec<LocalInfo>>

Calling this method will remove any pre-existing
IdentitySecureChannelLocalInfo entries from the given LocalInfo vector
before creating a new entry and appending it to the return vector.

This fix also removes the IdentitySecureChannelLocalInfo constructor
to ensure that future code has to use the new mark() method.

This closes #2868