refactor(rust): strategy to enable/disable logs in ockam_command
Previously, we had the arguments `verbose`, `quiet`, `spawned` + the OCKAM_LOG env variable to
control how logs were enabled. The OCKAM_LOG was the highest priority parameter, making it
impossible to override with any of mentioned command arguments.
Also, the `quiet` argument was ignored.

Now, the `spawned` argument has been removed and the priority has been reversed. That is, if
`quiet` is set, logs are not enabled. If `verbose` is set, its value will be used to define
the log level; if not, it will try to read the log level from the OCKAM_LOG env variable.
Logs won't be enabled if both `verbose` and OCKAM_LOG are not set.