refactor(rust): add sub-commands for portal and transport creation
This commit partially undoes the previous work towards addon
commands.  The main issue is with creating meaningful but dynamic help
pages if we leave clap's structure of parsing parameters to pass to
plugins (even if internal).

Comments and preparations are left in the code to use to upgrade this
system in the future.