feat(rust): use identity secure channels to communicate with orchestrator
As a workaround, we were passing the identity ID from client ->
orchestrator either in the path or in the request body. Now that
rust <-> elixir identity secure channels are working, switch to
properly use them and remove the workarounds.

Also attach the local_metadata into the API Request structs,
to access the caller' identity.