feat(rust): add `message-format` global arg
This argument is only used in one command as an example right now. In future PR's, this argument
will be used to control the format of the command output. This can be helpful to pipe commands,
processing the output in json format and select the variables we need as input for the next
command.

As part of this PR, a small refactor has been done to pass a `CommandGlobalOpts` object to
subcommands, which includes the global command arguments as well as the `OckamConfig` instance.
Ideally, all commands should receive this object, so we can handle how the `message-format`
applies to each command individually, for example.