feat(elixir): add kinesis init_partition callback
When initiating partition, we're calling AWS Kinesis with DescribeStream operation
to retrieve HashKey and StartingSequenceNumber for this particular partition (shard).
We need HashKey to control to which shard a message will be put
(otherwise it would get put to a shard based on hashed partition key).
StartingSequenceNumber is needed to identify first valid index in the shard.