feat(elixir): add kinesis save callback
When saving a message we're using a hash key previously stored in state
to ensure a message lands in the shard we want.
To ensure that the indexes (Kinesis sequence numbers) are always increasing
we use a sequence number of previously pushed message as a sequence_number_for_ordering reference.