fix(rust): `project create` command now works when services + node + cloud args are passed
When using `node_opts` and `cloud_opts` at the command level (project/mod.rs), it was
causing some weird behaviors when we tried to use those arguments mixed with Vec<T>
arguments. More specifically, clap is not able to parse the arguments in the expected order
making it impossible to pass both a Vec<T> argument and any of the global arguments defined
at `node_opts` and `cloud_opts`.

In this commit, the global arguments have been moved at the subcommand level so we can put
them *before* the Vec<T> arguments, and now clap is able to parse all arguments properly.