feat(rust): add `SHOW_HIDDEN` environment variable
We want to incremently ship commands that are ready. This commit
introduces a SHOW_HIDDEN environment variable and hides many commands
and options that are early in their development behind this variable.

These commands will still function, but they won't be shown in help
text. To see them run `SHOW_HIDDEN=true ockam ...`