feat(rust): make `IdentityIdentifier` encodable
Derive `minicbor::{Encode, Decode}` for `IdentityIdentifier` to enable
direct use of the type in API requests and responses.

Also changes `Identiy::identifier` to return `&IdentityIdentifier`
instead of `Result<IdentityIdentifier>` as the method never fails and
to better support callers that do not need to get a clone of the value.