feat(rust): derive `Default` for `Id` and `Error`
To allow construction of responses and errors without having access to
the request (e.g. because decoding it failed), this commit derives
`Default` for `Id` and ensures that random Ids are never 0. It also
makes the `path` in `Error` optional, so that an error response can be
constructed without knowing the request path. This functionality should
be used judiciously, e.g. only when request decoding failed, otherwise
when the data is available it should be used.