fix(rust): creation of static forwarder at local nodes
Adds a more elaborate logic to derive from the `--at` argument if it references a local node
or a remote node.

Also, the `address` argument has been removed because it was not being used and the `for_node`
argument has been renamed to `to`.

Finally, the forwarder service has been moved to its own module (no logic changes done,
just moved the `create_forwarder` function).