refactor(rust): make node delete command more robust when "force" argument is present
It largely improves error handling, depending less on `std::process::exit` and propagating errors
only when needed.

When using the `--force` argument, it will try to delete all the nodes present in the
config file, but won't propagate any error if it fails to delete a given node or if there are
no nodes in the config file. Also, it will always try to delete the node's PID and its
content from the config file, regardless of whether one of the two operations fails.

Additionally, when the `--all` argument is also passed, it will remove both the config
directory and the nodes directory.