feat(rust): add project info command
Previously the output of project show was given to new members that
enroll, however it contained private information (e.g. the users list)
which are of no concern to an individual new member. This PR adds a
project info structure that contains only the relevant information for
the enrolling member.