fix(rust): project enroll
Currently `project enroll` establishes a secure channel to the project
node instead of the project's authority. This commit extends the
`ProjectLookup` data to include the authority access route. `project
enroll` will attempt to map a `/project` protocol as given in the input
multi-address to the corresponding authority address and establish a
secure channel to the authority.

Example:

  ockam project enroll \
    --member ... \
    --to /project/myproject/service/authenticator