fix(rust): print node status will wait only when creating new nodes
The `node create|list|show` commands run a method that retrieves the node status. In a previous PR,
this method was modified to block until the node was ready. This introduced a deadlock issue when
running the `list` and `show` commands. This commit changes that behaviour so that the `create`
command is the only one waiting until the node is up.