fix(rust): set project data into the lookup table when creating a node
When creating a node with the `--project` argument, a json file was loaded and parsed as an
instance of `ProjectInfo` to set up the project authority. This object is now also used to
add the project data to the node's config lookup table, so the user can later use
`/project/<name>` routes. Additionally, `ProjectInfo` now includes the project's `name` and
`access_route`.