feat(rust): recover remote forwarder
Currently any interruption to the underlying TCP connection of a remote
forwarder will permanently make it unusable. To remedy the sitution a
session "medic" is added to the node manager that maintains sessions
which in turn provide actions to replace themselves. Sessions are each
pinged at regular frequency and upon three subsequent failures the
replacement action is triggered.

Currently only remote forwarders to API nodes, i.e. with addresses like

  /dnsaddr/hostname/tcp/12345/service/api
  /ip4/1.2.3.4/tcp/5678/service/api
  /ip6/::1/tcp/12345/service/api

are considered and only if an authorised identity is present. Other
addresses or adresses like these but without authorised identities will
not enable recovery and not even attenpt to open a secure channel.

To allow the node manager to re-establish a secure channel the command
layer no longer performs the establishment of the secure channel, but
instead only project and node references are resolved and the resulting
address is passed as is to the node manager. NB that this assumes that
the project route does not change which is not always guaranteed. This
requires making project resolution available in the manager (tbd).