fix(rust): generalize `extract_node_name` so it can correctly parse other protocols
The function `extract_node_name` was failing on the `tcp-outlet create` command when passing
the `--from /service/outlet` argument, as it only worked with the `node` protocol.

Now, the function has been renamed to `extract_address_value` and it can be used to extract
the address from `node`, `service` and `project` protocols.