chore(rust): allow to store custom attributes for issuing credentials
Up to now, we store just a boolean indicating that given identity
identifier is a _member_ of the project.  When issuing credentials,
this was checked and a role=member credential was issued. This change
allows to store attributes, paving the way to attach different attributes
and values to credentials issued by a given authority.
Code and test is added to ensure new code can interpret the legacy
format.