test(rust): add an object pool to run commands tests that depend on local nodes only
Adding small tests to the `commands.bats` file to test the commands is slowing down the CI
pipelines a lot. The core idea behind the proposed changes is to be able to run the tests
that depend on local state to rust-based tests and reuse a small set of nodes, so we can
avoid creating and destroying nodes before every test (like it's done in the commands.bats file),
which is the operation that takes most time to complete.

The object pool it's lazy initialized with up to 16 nodes, so it doesn't block on computers
with lots of cores. Another approach would be to limit the size to 3 for example and run
the tests with at most 2 threads, but that would also require changes in the CI files.