fix(rust): change if evaluation of abac expressions
Currently if expressions evaluate both, their true branch and their
false branch. This does not worj when using `exist?` to check if
identifiers are bound as the use of unbound identifiers in either
branch would cause an error, defeating the purpose of `exists?`. To fix
this, the evaluation of `if` expressions now only evaluates the test
predicate and only then, depending on the outcome, will either the true
or the false branch be evaluated.