refactor(rust): remove old config.json file and add migration
The `config.json` file in the root $OCKAM_HOME directory contained a map with data about
spaces and projects. This data was duplicated in the `projects` directory and the CLI
code was not handling these two different configurations correcly.

A new cli state has been added for spaces and a new migration is in place to correctly
move the space and project lookups into the new `spaces` and `projects` directories
within the CliState.