fix(elixir): cache identity used by healthcheck
Healthcheck is generating a new identity on every check, which means
potentially thousands of new identities per day.
Since it does not clean up the identities, keys data can fill up memory
or disk where they stored.
It's not necessary to use a new identity every time, so they can be cached.

Using :persistent_term to cache identity for healthchecks

Added check_local_private_key function to Ockam.Identity to verify that it's a local
identity and it can be used to create signatures