fix(rust): two foreground nodes with same name can't be running at the same time
The `node create <name> -f` command could be run as two different processes while
targeting the same node state/configuration. A new check has been added to avoid that.

Additionally, there was an inconsistency when creating a background node, where it
would return an error when running the command for a stopped node. This could be done
with foreground nodes.