fix(rust): store a single transport in node config to avoid misuse
These changes fix a problem where a foreground node forcibly killed (e.g. due to an OS restart)
was unable to get recreated, as it wasn't removing the previous tcp-listener.

To avoid this kind of misuse, the list of transports stored as part of the node config has been
replaced by a single transport, which will be overwritten upon restarting a node.

To provide more context:
- We were storing a list of transports so we could recreate them after a node restart, but this
feature was never developed
- We were using the first tcp transport of that list to enable communication between the command
 and the api, which felt unnecessary