feat(rust): add support for human-writable durations
This adds a function ockam_command::util::duration::duration_parser,
which is intended to be used as part of the clap arg decoration, as the
value_parser argument.

Note, I intentionally did not convert durations which are passed to
other ockam projects, such as the timeout in
project::addon:configure_okta.