feat(rust): option to enroll without creating space nor project
The purpose of this feature is add an option to the ockam enroll so that only the
first step is executed using the --user-account-only command line flag. If a current
Space and/or Project already exist, they will continue to be retrieved and set as
the user defaults. If they do not exist, they will not be created as part of the
enroll command.